Feature/refactor merged
* Refacto module-rlotties.js

* Add loading screen when downloading gif file

* Update shortcuts modal design

* Add loading screen when downloading gif file

* Update prettyview logo - navbar

* Fix prettyview logo - navbar.vue


* Refactor index.html

* Separate javascript files

* Refactor vue component

* Update buildfiles

* Feature/chaewon

* Add loading screen when downloading gif file

* Update shortcuts modal design

* Add loading screen when downloading gif file

* Fix export modal close button to be disabled while creating gif file

* Delete shortcuts button

* Add shortcuts button and grid design

* Add shortcuts button and grid design

* Update design
