+++ main.js
@@ -1,10 +1,18 @@
 (function () {
     // load the script depending on web worker support
-    var src;
+    var srcs;
     if (typeof(Worker) !== "undefined") {
-        srcs = ['module-rlottie.js', './gifjs/dist/gif.js', './gifjs/dist/gif.worker.js', 'thumbnail.js'];
+        srcs = [
+            'index.js', 
+            'layernode.js',
+            'rlottie-module.js',
+            'rlottie-handler.js',
+            'thumbnail-handler.js', 
+            './gifjs/dist/gif.js', 
+            './gifjs/dist/gif.worker.js',
+        ];
     } else {
-        srcs = ['module-rlottie.js', './gifjs/dist/gif.js', './gifjs/dist/gif.worker.js', 'thumbnail.js'];
+        srcs = [];
     }
     var head = document.head;
