+++ style.css
@@ -67,9 +67,6 @@ html, body {
   border-style: groove;
 }
 
-.color_text {
-}
-
 .color_button_group {
   display: flex;
   flex-direction: row;
@@ -92,42 +89,39 @@ html, body {
 .btn_green { background-color: green;}
 .btn_blue { background-color: blue;}
 
-.card {
+/* .card {
   #box-shadow: 0 19px 38px rgba(0,0,0,0.30), 0 15px 12px rgba(0,0,0,0.22);
-}
-
-
-// copied
+} */
 
 .justify-center {
-    justify-content: center;
+  justify-content: center;
 }
 
 .text-xs {
-    font-size: .75rem;
+  font-size: .75rem;
 }
 
 .border-grey-light {
-    border-color: #dae1e7;
+  border-color: #dae1e7;
 }
 
 .bg-grey-lighter {
-    background-color: #f1f5f8;
+  background-color: #f1f5f8;
 }
 
 .rounded {
-    border-radius: .25rem;
+  border-radius: .25rem;
 }
 
 .text-grey-darkest {
-    color: #3d4852;
+  color: #3d4852;
 }
 .font-bold {
-    font-weight: 700;
+  font-weight: 700;
 }
 
 .cursor-pointer {
-    cursor: pointer;
+  cursor: pointer;
 }
 
 .margin-left-10px {
@@ -141,7 +135,6 @@ html, body {
 
 .scroll-sect-dark::-webkit-scrollbar-track {
   background: #37474F;
-  /* border-radius: 15px; */
 }
 
 .scroll-sect-dark::-webkit-scrollbar-corner {
@@ -164,7 +157,6 @@ html, body {
 
 .scroll-sect-light::-webkit-scrollbar-track {
   background: #CFD8DC;
-  /* border-radius: 15px; */
 }
 
 .scroll-sect-light::-webkit-scrollbar-corner {
@@ -178,4 +170,76 @@ html, body {
 .scroll-sect-light::-webkit-scrollbar-button {
   background-color: red;
   height: 0;
+}
+
+.player {
+  height: 92vh;
+  padding: 0;
+  background-color: none;
+}
+
+.screen {
+  height: 68vh;
+  background-color: none
+}
+
+.toolbox-basic {
+  height: 10vh;
+  width: 100%;
+}
+
+.play-button {
+  background-color: none;
+}
+
+.player-slider {
+  min-width: 30vw;
+}
+
+.extra-tools {
+  height: 12vh;
+}
+
+.scroll-sect {
+  overflow-y: scroll; 
+  height: 10vh;
+}
+
+.common-canvas {
+  width: 100px;
+  height: 100px;
+  border: 1px solid #BEBEBE;
+  transition: width 0.7s, height 0.7s, border-width 0.7s, border-radius 0.7s; 
+  z-index: 1;
+}
+
+.multiview-resizing {
+  width: 100px;
+  height: 100px;
+  border: 1px solid #BEBEBE;
+  transition: width 0.7s, height 0.7s, border-width 0.7s, border-radius 0.7s; 
+  z-index: 1;
+}
+
+.selected-canvas {
+  outline: 5px solid #0fccce;
+  transition: width 0.7s, height 0.7s, border-width 0.7s, border-radius 0.7s; 
+  z-index: 1;
+}
+
+.zindex2 {
+  z-index: 2;
+}    
+
+.canvas-chip {
+  position: relative;
+  bottom: 14vh;
+}
+
+.zindex3 {
+  z-index: 3;
+}
+
+.snapshot-container {
+  background-color: white;
 }
