+++ thumbnail-handler.js
